Implement IClassFixture instead of ClassInitialize
